finished all crud operations/authorizations

+++ src/trips/trips.service.ts
@@ -2,7 +2,6 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult } from 'typeorm';
 
-import { GetUser } from '@Auth/get-user.decorator';
 import { Trip } from '@Auth/trip.entity';
 import { TripRepository } from '@Auth/trip.repository';
 
@@ -18,14 +17,16 @@ export class TripsService {
     private readonly tripRepository: TripRepository,
   ) { }
 
-  public async getAllTrips(): Promise<Trip[]> {
-    const found: Trip[] = await this.tripRepository.find();
+  public async getAllTrips(tuser: User): Promise<Trip[]> {
 
-    return found;
+    return this.tripRepository.getAllTrips(tuser);
   }
 
-  public async getTripById(id: number): Promise<Trip> {
-    const found: Trip = await this.tripRepository.findOne(id);
+  public async getTripById(
+    id: number,
+    tuser: User,
+  ): Promise<Trip> {
+    const found: Trip = await this.tripRepository.findOne({ where: { id, tuserId: tuser.id } });
 
     if (!found) {
       throw new NotFoundException(`Trip with ID ${id} was not found.`);
@@ -41,8 +42,11 @@ export class TripsService {
     return this.tripRepository.createTrip(createTripDto, tuser);
   }
 
-  public async deleteTrip(id: number): Promise<void> {
-    const result: DeleteResult = await this.tripRepository.delete(id);
+  public async deleteTrip(
+    id: number,
+    tuser: User,
+  ): Promise<void> {
+    const result: DeleteResult = await this.tripRepository.delete({ id, tuserId: tuser.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Trip with ID ${id} was not found.`);
